opdracht 6 + praktijkopdracht

+++ src/h06/Verdeling.java
@@ -5,8 +5,8 @@ import java.applet.*;
 
 public class Verdeling extends Applet {
 
-    int verdiend;
-    int verdeling;
+    double verdiend;
+    double verdeling;
 
 // DIT IS OPDRACHT 6.1
